added user profile finally

+++ app/RatTracker2k17/app/src/main/java/edu/gatech/cs2340/rattracker2k17/Data/Types.java
@@ -33,8 +33,8 @@ public class Types {
          * @param level the integer representation of the security level you are trying to obtain
          * @return the SecurityLevel enumerated value
          */
-        public static SecurityLevel parseSecurityLevel(int level) {
-            switch(level) {
+        public static SecurityLevel parseSecurityLevel(long level) {
+            switch((int)level) {
                 case 1: return Guest;
                 case 5: return User;
                 case 25: return Admin;
